Odyssey_Anime_Project/requirements.txt
python-telegram-bot
requests
APScheduler
